Extract AI review hook
